Commented out ui_chara_db editing
There are two reasons for this. First, the current implementation (where there is only one ui_chara_db for all your mods) expects you to continuously copy this file - which keeps getting edited - to your mod directory everytime you make an addition or removal. Second, the only fully-functional change within this file is the number of slots per character; for something that minor and self-explanatory, it's better to ask the user to edit it themselves when needed.